get/set ref pll returns char, get sweep status returns char, amplitude is int32

+++ api/include/wsa_lib.h
@@ -214,7 +214,7 @@ struct wsa_sweep_list {
 	int8_t trigger_enable;
 	int64_t trigger_start_freq;
 	int64_t trigger_stop_freq;
-	int64_t trigger_amplitude;
+	int32_t trigger_amplitude;
 	enum wsa_gain gain_rf;
 };
